requirements.txt
fastapi
uvicorn
python-multipart
requests
langchain-google-genai
langchain-chroma
langchain-community
chromadb
pydantic
python-dotenv